Majorly Refactor SDK.
The goal of this refactor was twofold:
1) To rely on as much built-in CMake functionality as possible, and
2) To increase flexibility for platform construction and layout.

The SDK scripts previously placed a considerable number of implicit
constraints on the layout of the source directories, and generally what
constitutes a target. An attempt here has been made to construct the
platforms in more of a library style; a platform is now any subdirectory
that calls platform(), and the usual CMake install logic is used to
control layout.

The examples were made into an external project that is invoked multiple
times; once per target. This both allows it to use simple add_executable
definitions and allows us to include the real CMakeLists.txt file for
all the examples with the SDK install.

Finally, the Clang config files are now automatically generated; we do
lose a bit of flexibility here, but it shouldn't be hard to get it back
if need be. I'd like to move most of the current functionality into the
compiler driver anyway, so no big loss.

This removes around 200 lines. We do lose the run-XXX simulator targets,
but everything else made it through.